requirements.txt
PyQt6>=6.6.0
matplotlib>=3.8.0
pandas>=2.1.0
pyodbc>=5.0.1
Pillow>=10.0.0
openpyxl>=3.1.0
